+++ modules/core/arrow-core-data/src/main/kotlin/arrow/core/MapK.kt
@@ -1,13 +1,6 @@
-package arrow.data
+package arrow.core
 
 import arrow.Kind
-import arrow.core.Eval
-import arrow.core.None
-import arrow.core.Option
-import arrow.core.Some
-import arrow.core.Tuple2
-import arrow.core.identity
-import arrow.core.iterateRight
 import arrow.higherkind
 import arrow.typeclasses.Applicative
 
@@ -91,4 +84,6 @@ fun <K, A, B> Map<K, A>.foldLeft(b: Map<K, B>, f: (Map<K, B>, Map.Entry<K, A>) -
 fun <K, A, B> Map<K, A>.foldRight(b: Map<K, B>, f: (Map.Entry<K, A>, Map<K, B>) -> Map<K, B>): Map<K, B> =
   this.entries.reversed().k().foldLeft(b) { x, y: Map.Entry<K, A> -> f(y, x) }
 
-fun <K, V> mapOf(vararg tuple: Tuple2<K, V>): MapK<K, V> = if (tuple.isNotEmpty()) tuple.map { it.a to it.b }.toMap().k() else emptyMap<K, V>().k()
+// TODO; check duplicate in TupleN
+// fun <K, V> mapOf(vararg tuple: Tuple2<K, V>): MapK<K, V> =
+//  if (tuple.isNotEmpty()) tuple.map { it.a to it.b }.toMap().k() else emptyMap<K, V>().k()
