+++ modules/core/arrow-core-data/src/test/kotlin/arrow/core/MapKTest.kt
@@ -1,15 +1,15 @@
-package arrow.data
+package arrow.core
 
 import arrow.Kind2
 import arrow.core.extensions.eq
 import arrow.core.extensions.hash
 import arrow.core.extensions.semigroup
-import arrow.data.extensions.mapk.eq.eq
-import arrow.data.extensions.mapk.functor.functor
-import arrow.data.extensions.mapk.hash.hash
-import arrow.data.extensions.mapk.monoid.monoid
-import arrow.data.extensions.mapk.show.show
-import arrow.data.extensions.mapk.traverse.traverse
+import arrow.core.extensions.mapk.eq.eq
+import arrow.core.extensions.mapk.functor.functor
+import arrow.core.extensions.mapk.hash.hash
+import arrow.core.extensions.mapk.monoid.monoid
+import arrow.core.extensions.mapk.show.show
+import arrow.core.extensions.mapk.traverse.traverse
 import arrow.test.UnitSpec
 import arrow.test.generators.mapK
 import arrow.test.laws.HashLaws
@@ -17,7 +17,6 @@ import arrow.test.laws.MonoidLaws
 import arrow.test.laws.SemigroupLaws
 import arrow.test.laws.ShowLaws
 import arrow.test.laws.TraverseLaws
-import arrow.test.laws.fix
 import arrow.typeclasses.Eq
 import io.kotlintest.properties.Gen
 import io.kotlintest.runner.junit4.KotlinTestRunner
