+++ modules/core/arrow-extras-data/src/main/kotlin/arrow/data/SortedMapK.kt
@@ -6,6 +6,7 @@ import arrow.core.Option
 import arrow.core.Tuple2
 import arrow.core.identity
 import arrow.core.iterateRight
+import arrow.core.k as mk
 import arrow.higherkind
 import arrow.typeclasses.Applicative
 
@@ -49,7 +50,7 @@ data class SortedMapK<A : Comparable<A>, B>(private val map: SortedMap<A, B>) : 
 
   fun <G, C> traverse(GA: Applicative<G>, f: (B) -> Kind<G, C>): Kind<G, SortedMapK<A, C>> = GA.run {
     map.iterator().iterateRight(Eval.always { just(sortedMapOf<A, C>().k()) }) { kv, lbuf ->
-      f(kv.value).map2Eval(lbuf) { (mapOf(kv.key to it.a).k() + it.b).toSortedMap().k() }
+      f(kv.value).map2Eval(lbuf) { (mapOf(kv.key to it.a).mk() + it.b).toSortedMap().k() }
     }.value()
   }
