+++ modules/core/arrow-extras-extensions/src/main/kotlin/arrow/data/extensions/setk.kt
@@ -6,7 +6,6 @@ import arrow.core.Tuple2
 import arrow.data.ForSetK
 import arrow.data.SetK
 import arrow.data.fix
-import arrow.data.k
 import arrow.extension
 import arrow.typeclasses.Eq
 import arrow.typeclasses.Foldable
