+++ modules/core/arrow-test/src/main/kotlin/arrow/test/generators/Generators.kt
@@ -26,7 +26,7 @@ import arrow.core.toOption
 import arrow.core.extensions.option.functor.map
 import arrow.core.extensions.option.functor.functor
 import arrow.core.ListK
-import arrow.data.MapK
+import arrow.core.MapK
 import arrow.core.NonEmptyList
 import arrow.data.SequenceK
 import arrow.data.SetK
