+++ modules/free/arrow-free-data/src/test/kotlin/arrow/free/Interpreters.kt
@@ -12,7 +12,7 @@ import arrow.core.ForListK
 import arrow.core.ForNonEmptyList
 import arrow.core.NonEmptyList
 import arrow.data.fix
-import arrow.data.k
+import arrow.core.k
 
 val cofreeOptionToNel: FunctionK<CofreePartialOf<ForOption>, ForNonEmptyList> = object : FunctionK<CofreePartialOf<ForOption>, ForNonEmptyList> {
   override fun <A> invoke(fa: Kind<CofreePartialOf<ForOption>, A>): Kind<ForNonEmptyList, A> =
