+++ modules/optics/arrow-optics/src/main/kotlin/arrow/optics/extensions/SetKAt.kt
@@ -1,7 +1,6 @@
 package arrow.optics.extensions
 
 import arrow.data.SetK
-import arrow.data.k
 import arrow.extension
 import arrow.optics.Lens
 import arrow.optics.PLens
