+++ modules/optics/arrow-optics/src/test/kotlin/arrow/optics/DSLTest.kt
@@ -2,15 +2,13 @@ package arrow.optics
 
 import arrow.core.None
 import arrow.core.ListK
-import arrow.data.MapK
-import arrow.data.k
-import arrow.optics.dsl.at
+import arrow.core.MapK
+import arrow.core.k
 import arrow.optics.extensions.listk.index.index
 import arrow.optics.extensions.mapk.at.at
 import arrow.optics.extensions.mapk.each.each
 import arrow.optics.extensions.traversal
 import arrow.test.UnitSpec
-import io.kotlintest.shouldBe
 
 @optics
 data class Street(val number: Int, val name: String) {
