+++ modules/optics/arrow-optics/src/test/kotlin/arrow/optics/instances/MapInstanceTest.kt
@@ -7,7 +7,7 @@ import arrow.core.extensions.option.eq.eq
 import arrow.core.extensions.option.monoid.monoid
 import arrow.core.extensions.semigroup
 import arrow.core.ListK
-import arrow.data.MapK
+import arrow.core.MapK
 import arrow.core.extensions.listk.eq.eq
 import arrow.optics.extensions.at
 import arrow.optics.extensions.each
